+++ src/lamp_py/ingestion/compress_gtfs/schedule_details.py
@@ -1,11 +1,8 @@
-import re
 import os
 import csv
 import zipfile
 import datetime
-import zoneinfo
 
-from urllib import request
 from typing import List
 from io import BytesIO
 from io import TextIOWrapper
@@ -15,22 +12,13 @@ from dataclasses import field
 import polars as pl
 
 from lamp_py.runtime_utils.process_logger import ProcessLogger
+from lamp_py.ingestion.utils import (
+    ordered_schedule_frame,
+    file_as_bytes_buf,
+)
+
 
 from .gtfs_schema_map import gtfs_schema
-
-
-def file_as_bytes_buf(file: str) -> BytesIO:
-    """
-    Create buffer of local or http(s) file path
-
-    :return BytesIO buffer
-    """
-    if file.startswith("http"):
-        with request.urlopen(file) as response:
-            return BytesIO(response.read())
-
-    with open(file, "rb") as f:
-        return BytesIO(f.read())
 
 
 # pylint: disable=R0902
@@ -195,107 +183,6 @@ class ScheduleDetails:
 
 
 # pylint: enable=R0902
-
-
-def date_from_feed_version(feed_version: str) -> datetime.datetime:
-    """
-    Extract date from feed_version text. Raise LookupError if no date found.
-
-    known date formats:
-        - YYYY-MM-DD
-        - MM/DD/YY
-    YYYY-MM-DD iso string will be converted from UTC to US/Eastern
-
-    :param feed_version: feed_version of gtfs schedule
-
-    :return: datetime extracted from feed_version text
-    """
-    utc_tz = datetime.timezone.utc
-    local_tz = zoneinfo.ZoneInfo("US/Eastern")
-
-    pattern_1 = r"(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})"
-    pattern_1_result = re.search(pattern_1, feed_version)
-
-    pattern_2 = r"(\d{1,2}\/\d{1,2}\/\d{2})"
-    pattern_2_result = re.search(pattern_2, feed_version)
-
-    if pattern_1_result is not None:
-        date_str = pattern_1_result.group(0)
-        date_dt = datetime.datetime.fromisoformat(date_str).replace(
-            tzinfo=utc_tz
-        )
-        date_dt = date_dt.astimezone(local_tz).replace(tzinfo=None)
-    elif pattern_2_result is not None:
-        date_str = pattern_2_result.group(0)
-        date_dt = datetime.datetime.strptime(date_str, "%m/%d/%y")
-    else:
-        raise LookupError(f"No date found in feed_version: '{feed_version}'")
-
-    return date_dt
-
-
-def ordered_schedule_frame() -> pl.DataFrame:
-    """
-    create de-duplicated and ordered frame of all MBTA gtfs schedules from
-    https://cdn.mbta.com/archive/archived_feeds.txt
-
-    de-duplicated on: published_date
-    ordered: oldest -> newest
-
-    :return frame with schema:
-    {
-        feed_start_date: int,
-        feed_version: str,
-        archive_url: str,
-        published_dt: datetime.datetime, (published date extracted from feed_version)
-        published_date: int, (published_dt date as YYYYMMDD int)
-    }
-    """
-    archive_url = "https://cdn.mbta.com/archive/archived_feeds.txt"
-    feed_columns = (
-        "feed_start_date",
-        "feed_version",
-        "archive_url",
-    )
-    feed_dtypes = {
-        "feed_start_date": pl.Int32,
-        "feed_version": pl.String,
-        "archive_url": pl.String,
-    }
-
-    # Accept-Encoding header required to avoid cloudfront cache-hit
-    req = request.Request(archive_url, headers={"Accept-Encoding": "gzip"})
-    with request.urlopen(req) as res:
-        feed = pl.read_csv(res.read(), columns=feed_columns, dtypes=feed_dtypes)
-
-    feed = (
-        feed.with_columns(
-            pl.col("feed_version")
-            .map_elements(date_from_feed_version, pl.Datetime)
-            .alias("published_dt"),
-        )
-        .with_columns(
-            pl.col("published_dt")
-            .dt.strftime("%Y%m%d")
-            .cast(pl.Int32)
-            .alias("published_date"),
-        )
-        .sort(
-            by=["feed_start_date", "published_dt"],
-        )
-        .unique(
-            subset="published_date",
-            keep="last",
-        )
-        .sort(
-            by="published_dt",
-        )
-    )
-
-    # fix_me: filter for malformed archive schedules
-    feed = feed.filter(feed["feed_start_date"] > 20180200)
-
-    return feed
 
 
 def schedules_to_compress(tmp_folder: str) -> pl.DataFrame:
