+++ tests/ingestion/test_gtfs_converter.py
@@ -359,14 +359,13 @@ def fixture_s3_patch(monkeypatch: MonkeyPatch) -> Iterator[None]:
         mock_write_parquet_file,
     )
 
-    def mock_move_s3_objects(files: List[str], to_bucket: str) -> List[str]:
-        """assert that all files moved go to archive and not error"""
-        if len(files) > 0:
-            assert "archive" in to_bucket
-        return []
+    def mock_gtfs_files_to_convert() -> List[str]:
+        """provide list of gtfs paths to convert"""
+        return [os.path.join(incoming_dir, "MBTA_GTFS.zip")]
 
     monkeypatch.setattr(
-        "lamp_py.ingestion.convert_gtfs.move_s3_objects", mock_move_s3_objects
+        "lamp_py.ingestion.convert_gtfs.gtfs_files_to_convert",
+        mock_gtfs_files_to_convert,
     )
 
     # everything before this yield is executed before running the test
@@ -396,8 +395,6 @@ def test_schedule_conversion(
     # pass in the test schedule file and convert. the monkey patched parquet
     # writer function will check that all tables generated by conversion were
     # processed correctly.
-    gtfs_schedule_file_list = [os.path.join(incoming_dir, "MBTA_GTFS.zip")]
-    converter.add_files(gtfs_schedule_file_list)
     converter.convert()
 
     # check through all of the paths that were "written" to s3 have the
